feat(coach): extract strategy preparation into dedicated feature
Moves StrategyPreparationZone and all related components/hook from
matches page components into src/features/coach/strategyPreparation/.
Adds season badge and recording_date desc sorting to CompactVideoList.
Fixes VideoWithMatch type to include seasons and corrects StrategyTabWithVideos prop type.
