+++ src/features/coach/index.ts
@@ -1,2 +1,3 @@
 export * from './layout';
 export * from './providers';
+export * from './strategyPreparation';
