+++ src/features/coach/strategyPreparation/components/HeadToHeadStatistics.tsx
