+++ src/features/coach/strategyPreparation/components/StrategyTabWithHeadToHead.tsx
