+++ src/features/coach/strategyPreparation/components/StrategyTabWithInfo.tsx
