+++ src/features/coach/strategyPreparation/components/index.ts
@@ -1,0 +1,8 @@
+export {default as StrategyPreparationZone} from './StrategyPreparationZone';
+export {default as CompactVideoList} from './CompactVideoList';
+export {default as TabWithVideos} from './StrategyTabWithVideos';
+export {default as TabWithStrategy} from './StrategyTabWithInfo';
+export {default as TabWithHeadToHead} from './StrategyTabWithHeadToHead';
+export {default as TabWithPreviousMatches} from './StrategyTabWithPreviousMatches';
+export {default as HeadToHeadStatistics} from './HeadToHeadStatistics';
+export {default as OpponentMatchStatistics} from './OpponentMatchStatistics';
