+++ src/features/coach/strategyPreparation/index.ts
@@ -1,0 +1,2 @@
+export {StrategyPreparationZone} from './components';
+export {useStrategyPreparation} from './hooks/useStrategyPreparation';
