+++ src/hooks/index.ts
@@ -2,7 +2,6 @@ export * from './admin/useExcelImport';
 export * from './admin/usePerformanceMonitoring';
 export * from './auth/useAuth';
 export {useAuth as useAuthNew} from './auth/useAuthNew';
-export * from './coach/useStrategyPreparation';
 export * from './coach/useUpcomingBirthdays';
 export * from './entities/attendance/data/useFetchAttendanceStatistics';
 export * from './entities/attendance/data/useFetchMembersAttendance';
